+++ raider/utils.py
@@ -131,20 +131,19 @@ def import_raider_objects() -> Dict[str, Any]:
 
     """
     hy_imports = {
-        "plugins": (
-            "Variable "
-            "Prompt "
+        "plugins.common": ("Empty " "Plugin " "Parser "),
+        "plugins.basic": (
             "Regex "
             "Html "
             "Json "
+            "Variable "
+            "Command "
+            "Prompt "
             "Cookie "
             "Header "
-            "Command "
-            "Plugin "
-            "Alter "
-            "Combine "
-            "Empty "
         ),
+        "plugins.modifiers": ("Alter " "Combine "),
+        "plugins.parsers": ("Parser " "UrlParser "),
         "flow": "Flow",
         "request": "Request PostBody Template",
         "operations": (